+++ fansly_downloader_ng.py
@@ -2,7 +2,9 @@
 
 """Fansly Downloader NG"""
 
-__version__ = '1.8.7'
+from app_version import APP_VERSION
+
+__version__ = APP_VERSION
 __date__ = '2025-12-06'
 __maintainer__ = 'prof79'
 __copyright__ = f'Copyright (C) 2023-2024 by {__maintainer__}'